Add latex system dependency and implement pdflatex rendering

+++ Sources/D2Utils/TemporaryDirectory.swift
@@ -1,20 +1,27 @@
 import Foundation
 
-/** A custom temporary directory. */
+/**
+ * A custom temporary directory. The directory is deleted
+ * when this instance is deinitialized.
+ */
 public class TemporaryDirectory {
-	private let url: URL
+	public let url: URL
+	public var deleteAutomatically: Bool = true
+	public var exists: Bool { return FileManager.default.fileExists(atPath: url.path) }
 	
 	public init() {
 		url = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
 	}
 	
-	public func childFile(name: String) -> TemporaryFile {
+	public func childFile(named name: String) -> TemporaryFile {
 		return TemporaryFile(url: url.appendingPathComponent(name))
 	}
 	
 	deinit {
 		do {
-			try FileManager.default.removeItem(at: url)
+			if deleteAutomatically && exists {
+				try FileManager.default.removeItem(at: url)
+			}
 		} catch {
 			print("Error while removing temporary directory: \(error)")
 		}
